+++ src/AntDesign.Charts/Components/Configs/ILegendMarkerStyle.cs
@@ -28,7 +28,7 @@ namespace AntDesign.Charts
         [JsonPropertyName("lineOpacity")]
         public int? LineOpacity { get; set; }
         [JsonPropertyName("opacity")]
-        public int? Opacity { get; set; }
+        public double? Opacity { get; set; }
         [JsonPropertyName("shadowColor")]
         public string ShadowColor { get; set; }
         [JsonPropertyName("shadowBlur")]
